Rename network referral cron from referrals/queue to referrals/network
