+++ apps/web/lib/partners/create-stablecoin-payout.ts
@@ -291,7 +291,7 @@ export const createStablecoinPayout = async ({
       enqueueBatchJobs(
         payoutIds.map((payoutId) => ({
           queueName: "create-referral-commissions",
-          url: `${APP_DOMAIN_WITH_NGROK}/api/cron/commissions/referrals/queue`,
+          url: `${APP_DOMAIN_WITH_NGROK}/api/cron/commissions/referrals/network`,
           body: {
             payoutId,
           },
