+++ apps/web/lib/partners/create-stripe-transfer.ts
@@ -258,7 +258,7 @@ export const createStripeTransfer = async ({
       enqueueBatchJobs(
         payoutIds.map((payoutId) => ({
           queueName: "create-referral-commissions",
-          url: `${APP_DOMAIN_WITH_NGROK}/api/cron/commissions/referrals/queue`,
+          url: `${APP_DOMAIN_WITH_NGROK}/api/cron/commissions/referrals/network`,
           body: {
             payoutId,
           },
